# Conflicts:
#	resources/views/member/profile.blade.php

+++ app/Http/Controllers/TalentNewController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Mail;
+use Maatwebsite\Excel\Concerns\WithHeadings;
 
 use App\Mail\UpscaleEmail;
 use App\Models\Talent;
@@ -153,7 +154,7 @@ class TalentNewController extends Controller
         // }
     }
 
-    public function export_excel()
+    public function export_excel ()
     {
         return Excel::download(new TalentExport, 'talent.xlsx');
     }
@@ -172,8 +173,11 @@ class TalentNewController extends Controller
             $talent = Talent::find($row);
             
             //menghapus semua data di table user yg berelasi
-            $user = User::find($talent->user_id); 
-            $user->delete() ; 
+            if ( $user = User::find($talent->user_id) ) 
+            {
+
+                $user->delete() ; 
+            } 
 
             //delete 
             Talent::where('talent_id', $row)->delete();
@@ -185,7 +189,7 @@ class TalentNewController extends Controller
         return view('admin.talent.insert');
     }
 
-public function insertData(Request $request){
+    public function insertData(Request $request){
 
 
 
@@ -282,7 +286,6 @@ public function insertData(Request $request){
         
 
         return redirect('admin/talent/list/insert')->with('success', 'Data Talent Berhasil dimasukkan.');
-
        
     }
